Application/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
torch
Pillow
requests
werkzeug
fastapi-cache
numpy